requirements.txt
numpy==1.23.4
pandas==1.5.1
Pillow==9.2.0
xlrd<2.0
scikit_learn==1.1.2
streamlit==1.13.0
